pausing in the middle to give a demo...
